refactor(process): use int for limitedBuffer capacity

+++ process/command.go
@@ -22,7 +22,7 @@ type Command struct {
 	Stdin io.Reader
 	// MaxOutputBytes bounds captured stdout and stderr independently.
 	// Zero or negative means unlimited capture.
-	MaxOutputBytes int64
+	MaxOutputBytes int
 	// GracePeriod is how long to wait after SIGTERM before SIGKILL.
 	// Defaults to 5 seconds if zero.
 	GracePeriod time.Duration
